refactor directory frontend
