+++ frontend/src/App.tsx
@@ -7,12 +7,12 @@ import {Container} from "@mui/material";
 import {BrowserRouter, Navigate, Route, Routes} from "react-router-dom";
 import {ToastContainer} from "react-toastify";
 import 'react-toastify/dist/ReactToastify.css';
-import RouteCollection from "./components/RouteCollection";
+import RouteCollection from "./components/routes/RouteCollection";
 import NavigationBottom from "./components/NavigationBottom";
-import RouteDetails from "./components/RouteDetails";
+import RouteDetails from "./components/routes/RouteDetails";
 import useCompareRoutes from "./hooks/useCompareRoutes";
-import CompareRoutesCollection from "./components/CompareRoutesCollection";
-import CompareRoutesDetails from "./components/CompareRoutesDetails";
+import CompareRoutesCollection from "./components/compare/CompareRoutesCollection";
+import CompareRoutesDetails from "./components/compare/CompareRoutesDetails";
 import LoginPage from "./components/LoginPage";
 import useUser from "./hooks/useUser";
 import ProtectedRoutes from "./components/ProtectedRoutes";
