+++ frontend/src/components/routes/RouteCard.tsx
@@ -2,8 +2,8 @@ import {Button, ButtonGroup, Card} from "@mui/material";
 import DeleteIcon from "@mui/icons-material/Delete";
 import {useNavigate} from "react-router-dom";
 import {useContext} from "react";
-import {RoutesContext} from "../contexts/RoutesContextProvider";
-import {Route} from "../models/RouteModel";
+import {RoutesContext} from "../../contexts/RoutesContextProvider";
+import {Route} from "../../models/RouteModel";
 import RouteVehicleIcon from "./RouteVehicleIcon";
 import RouteInfo from "./RouteInfo";
