+++ frontend/src/components/routes/RouteInfo.tsx
@@ -1,5 +1,5 @@
 import {Box, Divider, Typography} from "@mui/material";
-import {Route} from "../models/RouteModel";
+import {Route} from "../../models/RouteModel";
 
 type RouteInfoProps = {
     route: Route
